# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API specifications
  description: |
    API specifications of WASAPhoto, exam's project for
    the course Web and Software Architecture 2023/2024.
  version: 0.0.1
servers:
  - url: http://localhost:3000
...
components:
...
  securitySchemes:

# BearerAuthentication #
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string


#### SCHEMAS ####
  schemas:
...
    Image-path: 
      title: Image path
      description: |
        String with the path to the image.
      type: string
      format: uri
      pattern: "image[-]?[0-9]+\\.(jpg|png)"
      minLength: 1
      maxLength: 50
...
    User-ID:
      title: User ID
      description: User unique identifier.
      type: integer
      minimum: 1
      example: 1
      readOnly: true
    Username:
      title: Username
      description: |
        Name that the user has chosen, the user can be
        found by other user with the username, a user
        can change its username.
      type: string
      example: Mario
      pattern: "^.*?$"
      minLength: 3
      maxLength: 16

# User schema #
...
    User-summary:
      title: User summary
      description: |
        Summary of User object infos
      type: object
      properties:
        user-id:        { $ref: "#/components/schemas/User-ID" }
        username:       { $ref: "#/components/schemas/Username" }
        profile-image-path:  { $ref: "#/components/schemas/Image-path" }

### PHOTO RELATED SCHEMAS ###
...
paths:
...
  /users/{username}/followers/:
    parameters:
      - name: username
        in: path
        schema: { $ref: "#/components/schemas/Username" }
        required: true
    
    get:
      tags: ["User Operations"]
      summary: List the followers
      description: |
        List the followers of the specified user.
      
      operationId: listFollowers

      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: List of followers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User-summary" }
                minItems: 0
                maxItems: 1000000
        
        "400":
          description: Bad request

        "401":
          description: Unauthorized
        
        "404":
          description: User not found
        
        "500":
          description: Internal server error
...
  /users/{username}/banned/:
    parameters:
      - name: username
        in: path
        schema: { $ref: "#/components/schemas/Username" }
        required: true
    
    get:
      tags: ["User Operations"]
      summary: List the banned users
      description: |
        List the users banned by the specified user.
      
      operationId: listBanned

      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: List of banned users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User-summary" }
                minItems: 0
                maxItems: 1000000
        
        "400":
          description: Bad request

        "401":
          description: Unauthorized
        
        "404":
          description: User not found
        
        "500":
          description: Internal server error
...